refactoring take #1
